refactor: (#107) TeamGameResult 로직과 UI 분리
